+++ client-participation/js/strings/ja-JP.js
